@@ -3,43 +3,47 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/moment": "^1.3.13",
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-brands-svg-icons": "^5.12.1",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "@material-ui/core": "4.11.2",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.45",
-    "@material-ui/pickers": "^3.2.10",
+    "@babel/plugin-transform-class-properties": "^7.24.1",
+    "@babel/plugin-transform-nullish-coalescing-operator": "^7.24.1",
+    "@babel/plugin-transform-optional-chaining": "^7.24.1",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.0",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@monaco-editor/react": "^4.2.2",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/lab": "^5.0.0-alpha.169",
+    "@mui/material": "^5.15.14",
+    "@mui/styles": "^5.15.14",
+    "@mui/system": "^5.15.14",
+    "@mui/x-date-pickers": "^6.19.8",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "chart.js": "^3.9.1",
     "chartjs-adapter-moment": "^1.0.0",
     "country-list": "^2.2.0",
+    "date-fns": "^3.6.0",
     "emoji-mart": "^3.0.0",
     "faker": "^4.1.0",
     "google-palette": "^1.1.0",
     "immutable": "^4.0.0-rc.12",
     "lodash": "^4.17.19",
-    "material-ui": "^0.20.2",
     "moment": "^2.29.4",
     "monaco-editor": "^0.28.1",
-    "mui-datatables": "^3.7.4",
-    "react": "^16.12.0",
+    "mui-datatables": "^4.3.0",
+    "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.1",
     "react-chartjs-2": "^4.0.0",
     "react-code-blocks": "^0.1.4",
     "react-copy-to-clipboard": "^5.0.3",
-    "react-dom": "^16.12.0",
+    "react-dom": "^17.0.2",
     "react-router-dom": "^5.1.2",
-    "react-toastify": "^5.5.0",
-    "recompose": "^0.30.0"
+    "react-toastify": "^5.5.0"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "^4.0.0",
+    "@vitejs/plugin-react": "^4.2.1",
     "prettier": "^2.8.8",
     "vite": "^4.5.2"
   },
