@@ -1,4 +1,4 @@
-import { Grid, Typography } from "@material-ui/core";
+import { Grid, Typography } from "@mui/material";
 import PlayersHistory from "../PlayersHistory";
 import LogsHistory from "../LogsHistory";
 
