@@ -4,15 +4,15 @@ import {
     postData,
     showResponse,
 } from "../../utils/fetchUtils";
-import Dialog from "@material-ui/core/Dialog";
-import DialogActions from "@material-ui/core/DialogActions";
-import DialogContent from "@material-ui/core/DialogContent";
-import DialogTitle from "@material-ui/core/DialogTitle";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
 import { throttle } from "lodash/function";
-import Grid from "@material-ui/core/Grid";
-import TextField from "@material-ui/core/TextField";
+import Grid from "@mui/material/Grid";
+import TextField from "@mui/material/TextField";
 import React from "react";
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 import { Link as RouterLink } from "react-router-dom";
 
 const LoginModal = ({
