@@ -1,17 +1,10 @@
 import React, { useEffect, useRef, useState } from "react";
-import Badge from "@material-ui/core/Badge";
-import { Comment, Send } from "@material-ui/icons";
-import {
-  Box,
-  Button,
-  Chip,
-  Drawer,
-  Grid,
-  makeStyles,
-  TextField,
-} from "@material-ui/core";
-import IconButton from "@material-ui/core/IconButton";
-import Typography from "@material-ui/core/Typography";
+import Badge from "@mui/material/Badge";
+import { Comment, Send } from "@mui/icons-material";
+import { Box, Button, Chip, Drawer, Grid, TextField } from "@mui/material";
+import makeStyles from '@mui/styles/makeStyles';
+import IconButton from "@mui/material/IconButton";
+import Typography from "@mui/material/Typography";
 import moment from "moment";
 
 const useStyles = makeStyles((theme) => ({
@@ -39,7 +32,7 @@ const MessageHistory = ({ data }) => {
       <Box paddingLeft={2}>
         <Grid
           container
-          justify="flex-start"
+          justifyContent="flex-start"
           alignContent="flex-start"
           alignItems="flex-end"
           direction="column"
@@ -51,7 +44,7 @@ const MessageHistory = ({ data }) => {
               <Grid item key={index}>
                 <Grid
                   container
-                  justify="flex-start"
+                  justifyContent="flex-start"
                   alignContent="flex-start"
                   alignItems="flex-end"
                   direction="column"
@@ -60,13 +53,11 @@ const MessageHistory = ({ data }) => {
                     <Chip
                       style={{ height: "auto", paddingTop: "-10px" }}
                       color="primary"
-                      variant="default"
                       label={
                         <Typography align="left" className={classes.message}>
                           {message.reason}
                         </Typography>
-                      }
-                    />
+                      } />
                   </Grid>
                   <Grid item>
                     <Typography
