@@ -1,17 +1,17 @@
 import React from "react";
-import Grid from "@material-ui/core/Grid";
-import TextField from "@material-ui/core/TextField";
+import Grid from "@mui/material/Grid";
+import TextField from "@mui/material/TextField";
 import "react-toastify/dist/ReactToastify.css";
 import TextHistory from "../textHistory";
-import Autocomplete from "@material-ui/lab/Autocomplete";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-import Checkbox from "@material-ui/core/Checkbox";
+import Autocomplete from '@mui/material/Autocomplete';
+import FormControlLabel from "@mui/material/FormControlLabel";
+import Checkbox from "@mui/material/Checkbox";
 
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import FormControl from "@material-ui/core/FormControl";
-import Select from "@material-ui/core/Select";
+import InputLabel from "@mui/material/InputLabel";
+import MenuItem from "@mui/material/MenuItem";
+import FormHelperText from "@mui/material/FormHelperText";
+import FormControl from "@mui/material/FormControl";
+import Select from "@mui/material/Select";
 import { getSharedMessages } from "../../utils/fetchUtils";
 
 const Reason = ({
@@ -51,9 +51,6 @@ const Reason = ({
           <TextField
             {...params}
             label={label}
-            multiline
-            rows={1}
-            rowsMax={10}
             margin="dense"
             helperText={helperText}
           />
@@ -88,7 +85,7 @@ const TextInputBar = ({
   /* todo refactor */
   return (
     <Grid item xs={12}>
-      <Grid container justify="flex-start" direction="row" alignItems="center">
+      <Grid container justifyContent="flex-start" direction="row" alignItems="center">
         <Grid item xs={12} lg={3}>
           <FormControl>
             <InputLabel>Sort</InputLabel>
