@@ -1,13 +1,13 @@
-import { ButtonGroup, Button } from "@material-ui/core";
+import { ButtonGroup, Button } from "@mui/material";
 import React from "react";
-import FlagIcon from "@material-ui/icons/Flag";
-import Tooltip from "@material-ui/core/Tooltip";
-import StarIcon from "@material-ui/icons/Star";
-import StarBorder from "@material-ui/icons/StarBorder";
-import HowToRegIcon from "@material-ui/icons/HowToReg";
-import AccessTimeIcon from "@material-ui/icons/AccessTime";
-import BlockIcon from "@material-ui/icons/Block";
-import VisibilityIcon from "@material-ui/icons/Visibility";
+import FlagIcon from "@mui/icons-material/Flag";
+import Tooltip from "@mui/material/Tooltip";
+import StarIcon from "@mui/icons-material/Star";
+import StarBorder from "@mui/icons-material/StarBorder";
+import HowToRegIcon from "@mui/icons-material/HowToReg";
+import AccessTimeIcon from "@mui/icons-material/AccessTime";
+import BlockIcon from "@mui/icons-material/Block";
+import VisibilityIcon from "@mui/icons-material/Visibility";
 
 export const ActionButton = ({
   blacklisted,
