@@ -6,17 +6,16 @@ import {
   ListItemAvatar,
   ListItemText,
   Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 import { Link as RouterLink } from "react-router-dom";
 import React from "react";
 import { List, Map } from "immutable";
-import Tooltip from "@material-ui/core/Tooltip";
-import FileCopyIcon from "@material-ui/icons/FileCopy";
-import AnnouncementIcon from "@material-ui/icons/Announcement";
-import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
-import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp";
-import KeyboardArrowDownIcon from "@material-ui/icons/KeyboardArrowDown";
-import { pure } from "recompose";
+import Tooltip from "@mui/material/Tooltip";
+import FileCopyIcon from "@mui/icons-material/FileCopy";
+import AnnouncementIcon from "@mui/icons-material/Announcement";
+import ListItemSecondaryAction from "@mui/material/ListItemSecondaryAction";
+import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
+import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import { getName } from "country-list";
 import makePlayerProfileUrl from "../../../utils/makePlayerProfileUrl";
 
@@ -34,7 +33,7 @@ const getCountry = (country) => {
   );
 };
 
-export const PlayerHeader = pure(({ player }) => {
+export const PlayerHeader = ({ player }) => {
   const [showAll, setShowAll] = React.useState(false);
   const hasMultipleName = player.get("names") && player.get("names").size > 1;
 
@@ -182,4 +181,4 @@ export const PlayerHeader = pure(({ player }) => {
       </ListItemSecondaryAction>
     </ListItem>
   );
-});
+};
