@@ -1,6 +1,6 @@
-import { Grid } from "@material-ui/core";
+import { Grid } from "@mui/material";
 import React from "react";
-import Tooltip from "@material-ui/core/Tooltip";
+import Tooltip from "@mui/material/Tooltip";
 import moment from "moment";
 
 export const PlayerSighthings = ({ player }) => {
@@ -33,7 +33,7 @@ export const PlayerSighthings = ({ player }) => {
   return (
     <Grid
       container
-      justify="space-between"
+      justifyContent="space-between"
       spacing={0}
     >
       {vip_expiration ? (
