@@ -1,4 +1,4 @@
-import { Grid, GridListTile, makeStyles } from "@material-ui/core";
+import { Grid, ImageListItem } from "@mui/material";
 import React from "react";
 import { ActionButton } from "./ActionButton";
 import { PlayerHeader } from "./PlayerHeader";
@@ -6,40 +6,23 @@ import { PlayerFlags } from "./PlayerFlags";
 import { PlayerSighthings } from "./PlayerSighthings";
 import { PlayerPenalties } from "./PlayerPenalties";
 
-const useStyles = makeStyles((theme) => ({
-  paperTile: {
-    backgroundColor: theme.palette.background.paper,
-    minHeight: "100%",
-    padding: theme.spacing(2),
-  },
-  root: {
-    display: "flex",
-    flexWrap: "wrap",
-    justifyContent: "space-around",
-    overflow: "hidden",
-  },
-}));
-
 export default ({ onDeleteFlag, player }) => {
-  const classes = useStyles();
 
   return (
-    <GridListTile style={{ minHeight: "100%" }}>
+    <ImageListItem style={{ minHeight: "100%" }}>
       <Grid
         container
-        className={classes.paperTile}
         direction="column"
-        justify="space-between"
+        justifyContent="space-between"
       >
         <PlayerHeader  player={player} />
         <PlayerFlags
           player={player}
-          
           onDeleteFlag={onDeleteFlag}
         />
         <PlayerSighthings  player={player} />
         <PlayerPenalties  player={player} />
-        <Grid container justify="center">
+        <Grid container justifyContent="center">
           <Grid item>
             <ActionButton
               blacklisted={false}
@@ -58,6 +41,6 @@ export default ({ onDeleteFlag, player }) => {
           </Grid>
         </Grid>
       </Grid>
-    </GridListTile>
+    </ImageListItem>
   );
 };
