@@ -1,5 +1,5 @@
 import React from "react";
-import { withTheme } from "@material-ui/core/styles";
+import withTheme from '@mui/styles/withTheme';
 import {
   Button,
   Grid,
@@ -7,7 +7,7 @@ import {
   TextField,
   Typography,
   Tooltip,
-} from "@material-ui/core";
+} from "@mui/material";
 import {
   addPlayerToWatchList,
   get,
@@ -23,17 +23,17 @@ import Padlock from "../../components/SettingsView/padlock";
 import TextHistoryManager, { SelectNameSpace } from "./textHistoryManager";
 import TextHistory from "../textHistory";
 import ServicesList from "../Services";
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import FormControl from "@material-ui/core/FormControl";
-import Select from "@material-ui/core/Select";
+import InputLabel from "@mui/material/InputLabel";
+import MenuItem from "@mui/material/MenuItem";
+import FormHelperText from "@mui/material/FormHelperText";
+import FormControl from "@mui/material/FormControl";
+import Select from "@mui/material/Select";
 import { ManualPlayerInput, WordList } from "../commonComponent";
-import DeleteIcon from "@material-ui/icons/Delete";
-import AddIcon from "@material-ui/icons/Add";
-import SaveIcon from "@material-ui/icons/Save";
+import DeleteIcon from "@mui/icons-material/Delete";
+import AddIcon from "@mui/icons-material/Add";
+import SaveIcon from "@mui/icons-material/Save";
 import RealVip from "./realVip";
-import HelpIcon from "@material-ui/icons/Help";
+import HelpIcon from "@mui/icons-material/Help";
 import ServerName from "./serverName";
 import AutoSettings from "./autoSettings";
 
@@ -100,16 +100,10 @@ const Hook = ({
       <Grid item xs={2}>
         {actionType === "delete" ? (
           <React.Fragment>
-            <IconButton
-              edge="start"
-              onClick={() => onDeleteHook(myHook, myRoles)}
-            >
+            <IconButton edge="start" onClick={() => onDeleteHook(myHook, myRoles)} size="large">
               <DeleteIcon />
             </IconButton>
-            <IconButton
-              edge="start"
-              onClick={() => onUpdateHook(myHook, myRoles)}
-            >
+            <IconButton edge="start" onClick={() => onUpdateHook(myHook, myRoles)} size="large">
               <SaveIcon />
             </IconButton>
           </React.Fragment>
@@ -121,7 +115,7 @@ const Hook = ({
               setMyRoles("");
               setMyHook("");
             }}
-          >
+            size="large">
             <AddIcon />
           </IconButton>
         )}
@@ -472,7 +466,7 @@ class RconSettings extends React.Component {
           <Typography variant="h6">Automated broadcast cycle</Typography>
         </Grid>
         <Grid item xs={12}>
-          <Grid container justify="space-evenly">
+          <Grid container justifyContent="space-evenly">
             <Grid item>
               <Padlock
                 handleChange={(v) =>
@@ -506,8 +500,8 @@ class RconSettings extends React.Component {
             fullWidth
             label="Auto broadcast messages"
             multiline
-            rows={4}
-            rowsMax={30}
+            minRows={4}
+            maxRows={30}
             value={_.join(
               broadcastMessages.map((m) => m.replace(/\n/g, "\\n")),
               "\n"
@@ -543,7 +537,7 @@ class RconSettings extends React.Component {
           container
           spacing={1}
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
           
         >
@@ -571,8 +565,8 @@ class RconSettings extends React.Component {
             fullWidth
             label="Shared standard messages"
             multiline
-            rowsMax={30}
-            rows={4}
+            minRows={4}
+            maxRows={30}
             value={_.join(
               standardMessages.map((m) => m.replace(/\n/g, "\\n")),
               "\n"
@@ -594,23 +588,23 @@ class RconSettings extends React.Component {
             Save shared messages
           </Button>
         </Grid>
-        <Grid item  justify="center" xs={12}>
+        <Grid item  justifyContent="center" xs={12}>
           <Typography variant="h5">Blacklist player by Steam ID</Typography>
         </Grid>
-        <Grid item  justify="center" xs={12}>
+        <Grid item  justifyContent="center" xs={12}>
           <Blacklist  />
         </Grid>
-        <Grid item  justify="center" xs={12}>
+        <Grid item  justifyContent="center" xs={12}>
           <Typography variant="h5">Add player to watchlist</Typography>
         </Grid>
         <Grid item xs={12}>
           <ManualWatchList  />
         </Grid>
-        <Grid item  justify="center" xs={12}>
+        <Grid item  justifyContent="center" xs={12}>
           <Typography variant="h5">Manage services</Typography>
         </Grid>
-        <Grid item  justify="center" xs={12}>
-          <Grid container justify="center">
+        <Grid item  justifyContent="center" xs={12}>
+          <Grid container justifyContent="center">
             <Grid item md={8} xs={12}>
               <ServicesList  />
             </Grid>
@@ -639,7 +633,7 @@ class RconSettings extends React.Component {
         >
           <WebhooksConfig  type="camera" />
         </Grid> */}
-        <Grid item  justify="center" xs={12}>
+        <Grid item  justifyContent="center" xs={12}>
           <Typography variant="h5">
             Auto votekick toggle{" "}
             <Tooltip title="When enabled this feature manages the votekicks ingame by turning it off if the conditions you set below are met, and turning it back on if they are NOT met">
@@ -652,7 +646,7 @@ class RconSettings extends React.Component {
           container
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
           spacing={1}
         >
@@ -720,14 +714,14 @@ class RconSettings extends React.Component {
             />
           </Grid>
         </Grid>
-        <Grid item  justify="center" xs={12}>
+        <Grid item  justifyContent="center" xs={12}>
           <Typography variant="h5">Camera notification config</Typography>
         </Grid>
         <Grid
           container
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
         >
           <Padlock
@@ -749,7 +743,7 @@ class RconSettings extends React.Component {
           container
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
         >
           <Grid item xs={12}>
@@ -775,7 +769,7 @@ class RconSettings extends React.Component {
           container
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
         >
           <Grid item xs={12}>
@@ -795,7 +789,7 @@ class RconSettings extends React.Component {
           container
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
         >
           <Grid item xs={12}>
@@ -826,7 +820,7 @@ class RconSettings extends React.Component {
           xs={12}
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
         >
           <Button
@@ -842,7 +836,7 @@ class RconSettings extends React.Component {
           xs={12}
           
           alignContent="center"
-          justify="center"
+          justifyContent="center"
           alignItems="center"
         >
           <Button
