@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { get, postData, showResponse } from "../../utils/fetchUtils";
 import Editor from "@monaco-editor/react";
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 import { withRouter } from "react-router";
-import Grid from "@material-ui/core/Grid";
+import Grid from "@mui/material/Grid";
 import { CopyBlock, dracula } from "react-code-blocks";
 import { toast } from "react-toastify";
 
@@ -58,84 +58,81 @@ const UserSetting = ({
     }
   };
 
-  return (
-    <>
-      <Grid container spacing={2}>
-        <Grid item xs={12}>
-          <h1>{description}</h1>
-        </Grid>
-        <Grid item xs={12}>
-          <Button onClick={() => getData(getEndpoint)}>Refresh</Button>
-          <Button onClick={() => sendData(validateEndpoint)}>Validate</Button>
-          <Button onClick={() => sendData(setEndpoint)}>Save</Button>
-        </Grid>
-        <Grid
-          item
-          xs={12}
-          justify="center"
-          spacing={2}
-          alignItems="center"
-          justifyContent="center"
-          direction="column"
-        >
-          <Editor
-            width="90vw"
-            height="50vh"
-            defaultLanguage="json"
-            value={data}
-            onChange={setData}
-            defaultValue={""}
-            theme="vs-dark"
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <h2>errors</h2>
-          {errors.map((ele) => {
-            return ele.type === "InvalidConfigurationError" ? (
-              <div>
-                <div>
-                  <b>Missing Keys</b>: {ele.missing_keys.join(", ")}
-                </div>
-                <div>
-                  <b>Mandatory Keys</b>: {ele.mandatory_keys.join(", ")}
-                </div>
-                <div>
-                  <b>Provided Keys</b>: {ele.provided_keys.join(", ")}
-                </div>
-              </div>
-            ) : (
-              <div>
-                {Object.keys(ele).map((k) => {
-                  return (
-                    <div>
-                      <b>{k}</b>: {JSON.stringify(ele[k])}
-                    </div>
-                  );
-                })}
-              </div>
-            );
-          })}
-        </Grid>
-        <Grid item xs={12}>
-          <Button onClick={() => getData(getEndpoint)}>Refresh</Button>
-          <Button onClick={() => sendData(validateEndpoint)}>Validate</Button>
-          <Button onClick={() => sendData(setEndpoint)}>Save</Button>
-        </Grid>
+  return <>
+    <Grid container spacing={2}>
+      <Grid item xs={12}>
+        <h1>{description}</h1>
       </Grid>
       <Grid item xs={12}>
-        <h1>Documentation</h1>
-        <div style={{ textAlign: "left" }}>
-          <CopyBlock text={notes} language="json" wrapLines theme={dracula} />
-        </div>
+        <Button onClick={() => getData(getEndpoint)}>Refresh</Button>
+        <Button onClick={() => sendData(validateEndpoint)}>Validate</Button>
+        <Button onClick={() => sendData(setEndpoint)}>Save</Button>
+      </Grid>
+      <Grid
+        item
+        xs={12}
+        justifyContent="center"
+        spacing={2}
+        alignItems="center"
+        direction="column"
+      >
+        <Editor
+          width="90vw"
+          height="50vh"
+          defaultLanguage="json"
+          value={data}
+          onChange={setData}
+          defaultValue={""}
+          theme="vs-dark"
+        />
       </Grid>
       <Grid item xs={12}>
-        <h1>Model Schema</h1>
-        <div style={{ textAlign: "left" }}>
-          <CopyBlock text={schema} language="json" wrapLines theme={dracula} />
-        </div>
+        <h2>errors</h2>
+        {errors.map((ele) => {
+          return ele.type === "InvalidConfigurationError" ? (
+            <div>
+              <div>
+                <b>Missing Keys</b>: {ele.missing_keys.join(", ")}
+              </div>
+              <div>
+                <b>Mandatory Keys</b>: {ele.mandatory_keys.join(", ")}
+              </div>
+              <div>
+                <b>Provided Keys</b>: {ele.provided_keys.join(", ")}
+              </div>
+            </div>
+          ) : (
+            <div>
+              {Object.keys(ele).map((k) => {
+                return (
+                  <div>
+                    <b>{k}</b>: {JSON.stringify(ele[k])}
+                  </div>
+                );
+              })}
+            </div>
+          );
+        })}
       </Grid>
-    </>
-  );
+      <Grid item xs={12}>
+        <Button onClick={() => getData(getEndpoint)}>Refresh</Button>
+        <Button onClick={() => sendData(validateEndpoint)}>Validate</Button>
+        <Button onClick={() => sendData(setEndpoint)}>Save</Button>
+      </Grid>
+    </Grid>
+    <Grid item xs={12}>
+      <h1>Documentation</h1>
+      <div style={{ textAlign: "left" }}>
+        <CopyBlock text={notes} language="json" wrapLines theme={dracula} />
+      </div>
+    </Grid>
+    <Grid item xs={12}>
+      <h1>Model Schema</h1>
+      <div style={{ textAlign: "left" }}>
+        <CopyBlock text={schema} language="json" wrapLines theme={dracula} />
+      </div>
+    </Grid>
+  </>;
 };
 
 export default withRouter(UserSetting);
