@@ -1,5 +1,5 @@
 import React from "react";
-import LinearProgress from "@material-ui/core/LinearProgress";
+import LinearProgress from "@mui/material/LinearProgress";
 
 const AutoRefreshLine = ({
   intervalFunction,
