@@ -1,14 +1,14 @@
 import React from "react";
 import clsx from "clsx";
-import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import ExpandLessIcon from "@material-ui/icons/ExpandLess";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import ExpandLessIcon from "@mui/icons-material/ExpandLess";
 import {
   IconButton,
   Card,
   CardHeader,
   CardContent,
   Collapse,
-} from "@material-ui/core";
+} from "@mui/material";
 
 const CollapseCard = ({
   title,
@@ -38,7 +38,7 @@ const CollapseCard = ({
             onClick={handleExpandClick}
             aria-expanded={expanded}
             aria-label="show more"
-          >
+            size="large">
             {expanded ? <ExpandLessIcon /> : <ExpandMoreIcon />}
           </IconButton>
         }
