@@ -6,10 +6,10 @@ import {
   Grid,
   Popover,
   Tooltip,
-} from "@material-ui/core";
-import Chip from "@material-ui/core/Chip";
-import Autocomplete from "@material-ui/lab/Autocomplete";
-import TextField from "@material-ui/core/TextField";
+} from "@mui/material";
+import Chip from "@mui/material/Chip";
+import Autocomplete from '@mui/material/Autocomplete';
+import TextField from "@mui/material/TextField";
 
 export const WithPopver = ({ popoverContent, children }) => {
   const [anchorEl, setAnchorEl] = React.useState(null);
@@ -63,7 +63,7 @@ export const ManualPlayerInput = ({
   actionName,
   tooltipText,
 }) => (
-  <Grid container spacing={1} justify="space-between">
+  <Grid container spacing={1} justifyContent="space-between">
     <Grid item xs={6} md={3}>
       <TextField
         id="steam-id"
