@@ -1,11 +1,11 @@
 import React from "react";
-import { withStyles } from "@material-ui/core/styles";
-import MuiDialogTitle from "@material-ui/core/DialogTitle";
-import MuiDialogContent from "@material-ui/core/DialogContent";
-import MuiDialogActions from "@material-ui/core/DialogActions";
-import IconButton from "@material-ui/core/IconButton";
-import CloseIcon from "@material-ui/icons/Close";
-import Typography from "@material-ui/core/Typography";
+import withStyles from '@mui/styles/withStyles';
+import MuiDialogTitle from "@mui/material/DialogTitle";
+import MuiDialogContent from "@mui/material/DialogContent";
+import MuiDialogActions from "@mui/material/DialogActions";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
+import Typography from "@mui/material/Typography";
 
 const styles = (theme) => ({
   root: {
@@ -26,13 +26,10 @@ const styles = (theme) => ({
 const DialogTitle = withStyles(styles)((props) => {
   const { children, onClose, ...other } = props;
   return (
-    <MuiDialogTitle disableTypography {...other}>
+    <MuiDialogTitle {...other}>
       <Typography variant="h6">{children}</Typography>
       {onClose ? (
-        <IconButton
-          aria-label="close"
-          onClick={onClose}
-        >
+        <IconButton aria-label="close" onClick={onClose} size="large">
           <CloseIcon />
         </IconButton>
       ) : null}
